refactor(agent): deepen output normalization
